Put verification tag back - See commit description
The WorkFrame Google+ page/account is tied to my personal google account for some reason, thus trying to verify that the WorkFrame page owned the WorkFrame website resulted in Google spazzing out because Google webmaster will only stay logged in to my personal account.
I solved this by opening up the page settings for WorkFrame and going down the rabbit hole, navigating through the maze of one useless setting after another until I found a standalone username to manage WorkFrame's page. After setting up a password for it, I signed in as a normal Google account with that weird username and hoped that this would make the WorkFrame account agnostic to my personal account - which luckily succeeded.
Since this login session was not connected to my personal account, Google webmaster was forced to use the WorkFrame account and I wasted no time in setting up verification for the website.
Fuck you, Google. If you're going to integrate features, but half ass the implementation and not cover possible bugs, then we're fine with not having those features.